Separate correlation and histogram windows.
Revert the previous decision.

Also merge test_main_gui_ctrl and test_main_gui_windows into
test_main_gui
