+++ extra_foam/gui/icons/binning.png
