+++ extra_foam/gui/icons/correlation.png
